MA-23121-1 Fix build error in gralloc after enable cflags "-Werror=conversion"
After enable "-Werror=conversion" cflags in gralloc, the implicit conversion
will pop build error. Here fix these build error in gralloc. But the build
error in AOSP common code(hardware/interface) cannot fix directly. So we just
fix the implicit conversion issues in NXP gralloc code but not add the cflags
"-Werror=conversion" in Android.bp because the depending code in AOSP also have
similar issues.

Reviewed-on: http://androidsource.nxp.com/project/26487

+++ display/gralloc/aidl/NxpAllocator.cpp
@@ -91,20 +91,20 @@ ndk::ScopedAStatus NxpAllocator::allocate(const std::vector<uint8_t>& descriptor
     }
 
     std::vector<native_handle_t*> handles;
-    handles.resize(count, nullptr);
+    handles.resize(static_cast<size_t>(count), nullptr);
 
-    for (int32_t i = 0; i < count; i++) {
+    for (uint32_t i = 0; i < static_cast<uint32_t>(count); i++) {
         ndk::ScopedAStatus status = allocate(description, &outResult->stride, &handles[i]);
         if (!status.isOk()) {
-            for (int32_t j = 0; j < i; j++) {
+            for (uint32_t j = 0; j < i; j++) {
                 releaseBufferAndHandle(handles[j]);
             }
             return status;
         }
     }
 
-    outResult->buffers.resize(count);
-    for (int32_t i = 0; i < count; i++) {
+    outResult->buffers.resize(static_cast<size_t>(count));
+    for (uint32_t i = 0; i < static_cast<uint32_t>(count); i++) {
         auto handle = handles[i];
         outResult->buffers[i] = ::android::dupToAidl(handle);
         releaseBufferAndHandle(handle);
@@ -193,21 +193,21 @@ ndk::ScopedAStatus NxpAllocator::allocate2(const BufferDescriptorInfo& descripto
     BufferDescriptorInfoV4 descriptionV4 = convertAidlToIMapperV4Descriptor(descriptor);
 
     std::vector<native_handle_t*> handles;
-    handles.resize(count, nullptr);
+    handles.resize(static_cast<size_t>(count), nullptr);
 
-    for (int32_t i = 0; i < count; i++) {
+    for (uint32_t i = 0; i < static_cast<uint32_t>(count); i++) {
         ndk::ScopedAStatus status =
                 allocate(descriptionV4, &outResult->stride, &handles[i], initialDataspace);
         if (!status.isOk()) {
-            for (int32_t j = 0; j < i; j++) {
+            for (uint32_t j = 0; j < i; j++) {
                 releaseBufferAndHandle(handles[j]);
             }
             return status;
         }
     }
 
-    outResult->buffers.resize(count);
-    for (int32_t i = 0; i < count; i++) {
+    outResult->buffers.resize(static_cast<size_t>(count));
+    for (uint32_t i = 0; i < static_cast<uint32_t>(count); i++) {
         auto handle = handles[i];
         outResult->buffers[i] = ::android::dupToAidl(handle);
         releaseBufferAndHandle(handle);
